+++ CMakeLists.txt
@@ -167,6 +167,7 @@ list(APPEND NVFUSER_SRCS
   ${NVFUSER_SRCS_DIR}/index_compute.cpp
   ${NVFUSER_SRCS_DIR}/instrumentation.cpp
   ${NVFUSER_SRCS_DIR}/interval_analysis.cpp
+  ${NVFUSER_SRCS_DIR}/ir/allocation_utils.cpp
   ${NVFUSER_SRCS_DIR}/ir/base_nodes.cpp
   ${NVFUSER_SRCS_DIR}/ir/builder.cpp
   ${NVFUSER_SRCS_DIR}/ir/cloner.cpp
