+++ csrc/ir/utils.cpp
@@ -1685,125 +1685,4 @@ getReshapeInputAndOutputIds(TensorView* reshape_out_tv) {
   return std::make_pair(reshaped_root_ids, reshaped_logical_ids);
 }
 
-std::pair<std::optional<bool>, std::optional<bool>> splitContiguity(
-    const std::optional<bool>& contiguity) {
-  // Credits to @jacobhinkle:
-  // https://github.com/NVIDIA/Fuser/pull/1124#discussion_r1368682735
-  if (!contiguity.has_value()) {
-    return {std::nullopt, std::nullopt};
-  }
-  if (*contiguity) {
-    return {true, true};
-  } else {
-    return {true, false};
-  }
-}
-
-std::pair<bool, std::optional<bool>> mergeContiguity(
-    const bool outer_is_expanded,
-    const std::optional<bool>& outer_contiguity,
-    const bool inner_is_expanded,
-    const std::optional<bool>& inner_contiguity) {
-  // Statuses `b` and `e` are represented in the IR with isBroadcast() and
-  // hasExpandedExtent(). Status `C` means stops propagating because we know we
-  // can't alias at that point.
-  //
-  // o\i | t  f  b  e
-  // ----+-----------
-  //  t  | t  f  t  C
-  //  f  | C  C  f  C
-  //  b  | t  f  b  e
-  //  e  | C  C  e  e
-  if (!outer_contiguity.has_value() && !outer_is_expanded) {
-    return {true, inner_contiguity};
-  }
-  if (!inner_contiguity.has_value() && !inner_is_expanded) {
-    return {true, outer_contiguity};
-  }
-
-  // o\i | t  f  b  e
-  // ----+-----------
-  //  t  | t  f     C
-  //  f  | C  C     C
-  //  b  |
-  //  e  | C  C     e
-  if (outer_is_expanded && inner_is_expanded) {
-    return {true, std::nullopt};
-  }
-  if (outer_is_expanded || inner_is_expanded) {
-    return {false, std::nullopt};
-  }
-
-  // o\i | t  f  b  e
-  // ----+-----------
-  //  t  | t  f
-  //  f  | C  C
-  //  b  |
-  //  e  |
-  if (*outer_contiguity) {
-    return {true, inner_contiguity};
-  }
-  return {false, std::nullopt};
-}
-
-std::optional<LinkedHashMap<IterDomain*, std::optional<bool>>>
-canonicalizeAllocationToLogical(const TensorView* tv) {
-  const std::vector<IterDomain*>& logical = tv->getLogicalDomain();
-  const std::vector<IterDomain*>& allocation = tv->getMaybeAllocationDomain();
-  LinkedHashMap<IterDomain*, std::optional<bool>> allocation_to_contiguity;
-  for (auto&& [alloc_id, contiguity] : zip(allocation, tv->getContiguity())) {
-    allocation_to_contiguity.pushBack(alloc_id, contiguity);
-  }
-
-  for (Expr* transform : DependencyCheck::getAllExprsBetween(
-                             {logical.begin(), logical.end()},
-                             {allocation.begin(), allocation.end()}) |
-           std::views::reverse) {
-    auto* split = dynamic_cast<Split*>(transform);
-    if (split == nullptr) {
-      // We can handle merges using a similar logic if/when we need to.
-      return std::nullopt;
-    }
-
-    // When split->outer() is parallelized and split->inner() is serial, we
-    // remove split->outer() regardless of its position and replace
-    // split->inner() with split->in(). This way, even when split->outer() is
-    // not adjacent to split->inner() (e.g. when it's outermost), we can
-    // still undo the split.
-    //
-    // Several other cases that I haven't implemented for simplicity.
-    //
-    // When split->outer() is serial and split->inner() is parallelized, we
-    // could remove split->inner() and replace split->outer() with
-    // split->in() regardless of split->inner()'s position.
-    //
-    // When split->outer() and split->inner() are both parallelized, we could
-    // replace either of them with split->in() and remove the other.
-    NVF_ERROR(!split->inner()->isParallelized());
-
-    const auto [outer_contiguity, next_i] =
-        allocation_to_contiguity.erase(split->outer());
-    if (!split->outer()->isParallelized()) {
-      // Check adjacency only if split->outer() is not parallelized.
-      if (next_i == allocation_to_contiguity.end() ||
-          next_i->first != split->inner()) {
-        return std::nullopt;
-      }
-    }
-    const auto [inner_contiguity, merge_i] =
-        allocation_to_contiguity.erase(split->inner());
-    const auto [mergeable, contiguity] = mergeContiguity(
-        split->outer()->hasExpandedExtent(),
-        outer_contiguity,
-        split->inner()->hasExpandedExtent(),
-        inner_contiguity);
-    if (!mergeable) {
-      return std::nullopt;
-    }
-    allocation_to_contiguity.insert(merge_i, split->in(), contiguity);
-  }
-
-  return allocation_to_contiguity;
-}
-
 } // namespace nvfuser::ir_utils
